Extract mergeAdditionalPayOnyxData into a leaf module to break import cycles
